refactor modal into components
