+++ src/client/src/views/DashboardScreen.vue
@@ -10,18 +10,29 @@
       <div class="d-block text-center">
         <h1 class="display-1">Dashboard</h1>
       </div>
-      <b-container class="">
+      <button
+        v-for="board in BoardViews"
+        v-bind:key="board"
+        v-bind:class="['board-button', { active: currentBoardView === board }]"
+        v-on:click="currentBoardView = board"
+      >
+        {{ board }}
+      </button>
+      <BasicBoardComponent
+        v-show="currentBoardComponent === 'BasicBoardComponent'"
+        v-bind:sensorData="newSensorData"
+        class="board"
+      >
+      </BasicBoardComponent>
+      <DetailedBoardComponent
+        v-show="currentBoardComponent === 'DetailedBoardComponent'"
+        v-bind:sensorData="newSensorData"
+        class="board"
+      >
+      </DetailedBoardComponent>
+      <b-container>
         <b-row align-v="stretch">
-          <b-col>
-            <b-row align-v="baseline">
-              <b-col style="font-size: 2rem">
-                <b-table stacked hover :items="currentExecution"></b-table>
-              </b-col>
-              <div class="w-100"></div>
-              <b-col style="font-size: 2rem">
-                <b-table hover :items="currentSensorData"></b-table>
-              </b-col> </b-row
-          ></b-col>
+          <b-col> </b-col>
           <b-col>
             <div id="canvasParent" class="w-100 h-100">
               <div id="canvasChild">
@@ -39,135 +50,12 @@
         <b-container>
           <b-row>
             <b-col>
-              <b-button
-                class="mt-3"
-                variant="success"
-                block
-                @click="showStatsModal"
-                >Detailed View</b-button
-              ></b-col
-            >
-            <b-col>
               <b-button class="mt-3" variant="danger" block @click="hideModal"
                 >Close Me</b-button
               ></b-col
             >
           </b-row>
         </b-container>
-      </div>
-    </b-modal>
-    <b-modal
-      id="statsModal"
-      size="xl"
-      ref="raw-stats-modal"
-      hide-footer
-      :hide-header-close="true"
-    >
-      <b-container>
-        <b-row>
-          <b-col>
-            <!-- Panel div start -->
-            <div class="panel panel-primary">
-              <div class="panel-heading text-center">
-                <h3 class="panel-title">Color Intensity</h3>
-              </div>
-              <div class="panel-body">
-                <!-- Chart container -->
-                <div id="chart_container">
-                  <div id="cy_axis"></div>
-                  <div id="demo_chart" ref="colorpanel"></div>
-                </div>
-                <!-- End of chart container -->
-              </div>
-              <div class="panel-footer">
-                <p v-if="displayedColorValues.length > 0">
-                  <small>
-                    <span v-bind:style="{ color: dvColors.Red }"
-                      >Red: {{ displayedColorValues[0].Red }}
-                    </span>
-
-                    <span v-bind:style="{ color: dvColors.Green }">
-                      Green: {{ displayedColorValues[0].Green }}
-                    </span>
-
-                    <span v-bind:style="{ color: dvColors.Blue }">
-                      Blue: {{ displayedColorValues[0].Blue }}
-                    </span>
-                  </small>
-                </p>
-              </div>
-            </div>
-            <!-- Panel div end --></b-col
-          >
-          <b-col>
-            <!-- Panel div start -->
-            <div class="panel panel-primary">
-              <div class="panel-heading text-center">
-                <h3 class="panel-title">Obstacle Distance</h3>
-              </div>
-              <div class="panel-body">
-                <!-- Chart container -->
-                <div id="chart_container">
-                  <div id="oy_axis"></div>
-                  <div id="demo_chart" ref="obstaclepanel"></div>
-                </div>
-                <!-- End of chart container -->
-              </div>
-              <div class="panel-footer">
-                <p v-if="displayedObstacleValues.length > 0">
-                  <small>
-                    <span v-bind:style="{ color: dvColors.Obstacle }"
-                      >Distance:
-                      {{ displayedObstacleValues[0].ObstacleDistance }} cm
-                    </span>
-                  </small>
-                </p>
-              </div>
-            </div>
-            <!-- Panel div end --></b-col
-          >
-          <div class="w-100"></div>
-          <b-col
-            ><!-- Panel div start -->
-            <div class="panel panel-primary">
-              <div class="panel-heading text-center">
-                <h3 class="panel-title">Speed</h3>
-              </div>
-              <div class="panel-body">
-                <!-- Chart container -->
-                <div id="chart_container">
-                  <div id="sy_axis"></div>
-                  <div id="demo_chart" ref="speedpanel"></div>
-                </div>
-                <!-- End of chart container -->
-              </div>
-              <div class="panel-footer">
-                <p v-if="displayedSpeedValues.length > 0">
-                  <small>
-                    <span v-bind:style="{ color: dvColors.SpeedLeft }"
-                      >Left: {{ displayedSpeedValues[0].SpeedRight }} RPM
-                    </span>
-
-                    <span v-bind:style="{ color: dvColors.SpeedRight }">
-                      Right: {{ displayedSpeedValues[0].SpeedLeft }} RPM
-                    </span>
-                  </small>
-                </p>
-              </div>
-            </div>
-            <!-- Panel div end --></b-col
-          >
-          <b-col><p>Placeholder 4 of 4</p></b-col>
-        </b-row>
-      </b-container>
-      <div class="d-block text-center">
-        <b-button
-          class="mt-3"
-          variant="outline-danger"
-          block
-          @click="hideStatsModal"
-          >Close Me</b-button
-        >
       </div>
     </b-modal>
     <b-modal ref="error-modal" hide-footer :title.sync="alertTitle">
@@ -190,16 +78,21 @@
 <script>
 import { localhost } from "@/config/config.js";
 import axios from "axios";
-import Rickshaw from "rickshaw";
-import "rickshaw/rickshaw.min.css";
-let colorChart;
-let obstacleChart;
-let speedChart;
+
+import BasicBoardComponent from "@/components/BasicBoardComponent.vue";
+import DetailedBoardComponent from "@/components/DetailedBoardComponent.vue";
+
 export default {
+  components: {
+    BasicBoardComponent,
+    DetailedBoardComponent,
+  },
   data() {
     return {
       alertMessage: "Something went wrong",
       alertTitle: "Watch out!",
+      currentBoardView: "Basic",
+      BoardViews: ["Basic", "Detailed"],
       currentExecution: [
         {
           executing: "Left",
@@ -209,25 +102,6 @@ export default {
       currentSensorData: [],
       newSensorData: {},
       interval: null,
-      //graph stuff,
-      colorSeries: [],
-      obstacleSeries: [],
-      speedSeries: [],
-      renderEveryNth: 1,
-      updateInterval: 1,
-      streamFrequency: 1,
-      messageIndex: 0,
-      displayedColorValues: [],
-      displayedObstacleValues: [],
-      displayedSpeedValues: [],
-      dvColors: {
-        Red: "#cb503a",
-        Green: "#72c039",
-        Blue: "#0047AB",
-        Obstacle: "5C5553",
-        SpeedLeft: "#cb503a",
-        SpeedRight: "#72c039",
-      },
       c: null,
       ctx: null,
       vueCanvas: null,
@@ -240,6 +114,11 @@ export default {
       smallCar: null,
     };
   },
+  computed: {
+    currentBoardComponent: function () {
+      return this.currentBoardView + "BoardComponent";
+    },
+  },
   created() {
     window.addEventListener("unload", function () {
       sessionStorage.removeItem("mapData");
@@ -251,11 +130,6 @@ export default {
   },
   mounted() {
     this.$root.$on("bv::modal::shown", (bvEvent, modalId) => {
-      //console.log("Modal is about to be shown", bvEvent, modalId);
-      if (modalId === "statsModal") {
-        this.initChart();
-        this.colorReadyToRender = true;
-      }
       if (modalId === "dashboardModal") {
         this.c = document.getElementById("map");
         this.ctx = this.c.getContext("2d");
@@ -300,47 +174,7 @@ export default {
     hideErrorModal() {
       this.$refs["error-modal"].hide();
     },
-    showStatsModal() {
-      this.$refs["raw-stats-modal"].show();
-    },
-    hideStatsModal() {
-      this.$refs["raw-stats-modal"].hide();
-    },
-    parseSensorData() {
-      const tmpStore = [];
-      const currentSensorData = this.newSensorData;
-      const color = {
-        r: "Red",
-        g: "Green",
-        b: "Blue",
-      };
-      const rowColor = {
-        r: "danger",
-        g: "success",
-        b: "primary",
-      };
-      let dangerTrue = "";
-      if (currentSensorData.ObstacleDistance < 20) {
-        dangerTrue = "danger";
-      }
-      tmpStore.push(
-        {
-          type: "Speed",
-          data: `${currentSensorData.SpeedLeft} rpm ⚡ ${currentSensorData.SpeedRight} rpm`,
-        },
-        {
-          type: "Obstacle",
-          data: `${currentSensorData.ObstacleDistance} cm`,
-          _rowVariant: dangerTrue,
-        },
-        {
-          type: "Color",
-          data: color[currentSensorData.Color],
-          _rowVariant: rowColor[currentSensorData.Color],
-        }
-      );
-      this.currentSensorData = tmpStore;
-    },
+
     getSensorData() {
       const token = sessionStorage.getItem("token");
       axios.defaults.headers.common["x-access-token"] = token;
@@ -370,240 +204,16 @@ export default {
           );
           tmpData.time = Date.now();
           this.newSensorData = tmpData;
-          this.parseSensorData();
-          // Graph
-          this.updateDisplayedColorValues();
-          this.updateDisplayedObstacleValues();
-          this.updateDisplayedSpeedValues();
-          if (this.colorSeries.length < this.renderEveryNth) {
-            this.colorSeries.push({
-              Red: tmpData.Red,
-              Green: tmpData.Green,
-              Blue: tmpData.Blue,
-            });
-          }
-          if (this.obstacleSeries.length < this.renderEveryNth) {
-            this.obstacleSeries.push({
-              ObstacleDistance: tmpData.ObstacleDistance,
-            });
-          }
-          if (this.speedSeries.length < this.renderEveryNth) {
-            this.speedSeries.push({
-              SpeedLeft: tmpData.SpeedLeft,
-              SpeedRight: tmpData.SpeedRight,
-            });
-          }
-          /* Render-time! */
-          if (
-            this.colorSeries.length == this.renderEveryNth &&
-            this.colorReadyToRender
-          ) {
-            this.insertColorDatapoints(this.colorSeries, colorChart);
-            this.insertObstacleDatapoints(this.obstacleSeries, obstacleChart);
-            this.insertSpeedDatapoints(this.speedSeries, speedChart);
-            this.colorSeries = [];
-            this.obstacleSeries = [];
-            this.speedSeries = [];
-          }
         })
         .catch((error) => {
           this.alertTitle = "Error";
           this.alertMessage = error.response.data.message;
           this.hideModal();
-          this.hideStatsModal();
           this.showErrorModal();
           console.log(error.response.data.message);
         });
     },
-    /* Rickshaw.js initialization */
-    initChart() {
-      colorChart = new Rickshaw.Graph({
-        element: this.$refs.colorpanel,
-        width: "500",
-        height: "300",
-        renderer: "line",
-        min: 0,
-        max: 800,
-        series: new Rickshaw.Series.FixedDuration(
-          [
-            {
-              name: "Red",
-              color: "#EC644B",
-            },
-            {
-              name: "Blue",
-              color: "#446CB3",
-            },
-            {
-              name: "Green",
-              color: "#44B355",
-            },
-          ],
-          undefined,
-          {
-            timeInterval: this.updateInterval,
-            maxDataPoints: 100,
-            timeBase: new Date().getTime() / 1000,
-          }
-        ),
-      });
-      new Rickshaw.Graph.Axis.Y({
-        graph: colorChart,
-        orientation: "left",
-        tickFormat: function (y) {
-          return y.toFixed(1);
-        },
-        ticks: 5,
-        element: document.getElementById("cy_axis"),
-      });
 
-      obstacleChart = new Rickshaw.Graph({
-        element: this.$refs.obstaclepanel,
-        width: "500",
-        height: "300",
-        renderer: "line",
-        min: 0,
-        max: 200,
-        series: new Rickshaw.Series.FixedDuration(
-          [
-            {
-              name: "One",
-              color: "#EC644B",
-            },
-          ],
-          undefined,
-          {
-            timeInterval: this.updateInterval,
-            maxDataPoints: 100,
-            timeBase: new Date().getTime() / 1000,
-          }
-        ),
-      });
-      new Rickshaw.Graph.Axis.Y({
-        graph: obstacleChart,
-        orientation: "left",
-        tickFormat: function (y) {
-          return y.toFixed(1);
-        },
-        ticks: 5,
-        element: document.getElementById("oy_axis"),
-      });
-      speedChart = new Rickshaw.Graph({
-        element: this.$refs.speedpanel,
-        width: "500",
-        height: "300",
-        renderer: "line",
-        min: 0,
-        max: 60,
-        series: new Rickshaw.Series.FixedDuration(
-          [
-            {
-              name: "Left",
-              color: "#EC644B",
-            },
-            {
-              name: "Right",
-              color: "#446CB3",
-            },
-          ],
-          undefined,
-          {
-            timeInterval: this.updateInterval,
-            maxDataPoints: 100,
-            timeBase: new Date().getTime() / 1000,
-          }
-        ),
-      });
-      new Rickshaw.Graph.Axis.Y({
-        graph: speedChart,
-        orientation: "left",
-        tickFormat: function (y) {
-          return y.toFixed(1);
-        },
-        ticks: 5,
-        element: document.getElementById("sy_axis"),
-      });
-      this.resizeChart(colorChart);
-      this.resizeChart(obstacleChart);
-      this.resizeChart(speedChart);
-
-      window.addEventListener("resize", () => {
-        this.resizeChart(colorChart);
-        this.resizeChart(obstacleChart);
-        this.resizeChart(speedChart);
-      });
-    },
-    resizeChart(chart) {
-      chart.configure({
-        width: this.$refs.colorpanel.clientWidth,
-      });
-      chart.render();
-    },
-    /* Insert received datapoints into the chart */
-    insertColorDatapoints(messages, chart) {
-      for (let i = 0; i < messages.length; i++) {
-        let colorData = {
-          Red: messages[i].Red,
-          Green: messages[i].Green,
-          Blue: messages[i].Blue,
-        };
-        chart.series.addData(colorData);
-      }
-      chart.render();
-    },
-    insertObstacleDatapoints(messages, chart) {
-      for (let i = 0; i < messages.length; i++) {
-        let obstacleDistanceData = {
-          ObstacleDistance: messages[i].ObstacleDistance,
-        };
-        chart.series.addData(obstacleDistanceData);
-      }
-      chart.render();
-    },
-    insertSpeedDatapoints(messages, chart) {
-      for (let i = 0; i < messages.length; i++) {
-        let speedData = {
-          SpeedLeft: messages[i].SpeedLeft,
-          SpeedRight: messages[i].SpeedRight,
-        };
-        chart.series.addData(speedData);
-      }
-      chart.render();
-    },
-    /* Update displayed values every second on average */
-    updateDisplayedColorValues() {
-      if (this.messageIndex == this.streamFrequency) {
-        this.messageIndex = 0;
-        this.displayedColorValues = this.colorSeries;
-      } else if (this.messageIndex == 0) {
-        this.displayedColorValues = this.colorSeries;
-        this.messageIndex++;
-      } else {
-        this.messageIndex++;
-      }
-    },
-    updateDisplayedObstacleValues() {
-      if (this.messageIndex == this.streamFrequency) {
-        this.messageIndex = 0;
-        this.displayedObstacleValues = this.obstacleSeries;
-      } else if (this.messageIndex == 0) {
-        this.displayedObstacleValues = this.obstacleSeries;
-        this.messageIndex++;
-      } else {
-        this.messageIndex++;
-      }
-    },
-    updateDisplayedSpeedValues() {
-      if (this.messageIndex == this.streamFrequency) {
-        this.messageIndex = 0;
-        this.displayedSpeedValues = this.speedSeries;
-      } else if (this.messageIndex == 0) {
-        this.displayedSpeedValues = this.speedSeries;
-        this.messageIndex++;
-      } else {
-        this.messageIndex++;
-      }
-    },
     resizeCanvas() {
       this.c.style.width = "100%";
       this.c.style.height = "100%";
